enh: move duplicate code to mixin
