+++ src/shared/components/ncContextResource/ResourceSharees.vue
@@ -19,14 +19,10 @@
 </template>
 
 <script>
-import { generateOcsUrl } from '@nextcloud/router'
-import { getCurrentUser } from '@nextcloud/auth'
-import axios from '@nextcloud/axios'
-import debounce from 'debounce'
 import { NcSelect } from '@nextcloud/vue'
 import formatting from '../../../shared/mixins/formatting.js'
+import searchUserGroup from '../../../shared/mixins/searchUserGroup.js'
 import ShareTypes from '../../mixins/shareTypesMixin.js'
-import { showError } from '@nextcloud/dialogs'
 import '@nextcloud/dialogs/style.css'
 
 export default {
@@ -35,7 +31,7 @@ export default {
 		NcSelect,
 	},
 
-	mixins: [ShareTypes, formatting],
+	mixins: [ShareTypes, formatting, searchUserGroup],
 
 	props: {
 		receivers: {
@@ -54,12 +50,7 @@ export default {
 
 	data() {
 		return {
-			query: '',
 			value: '',
-			loading: false,
-			minSearchStringLength: 1,
-			maxAutocompleteResults: 20,
-			suggestions: [],
 			preExistingSharees: [...this.receivers],
 			localSharees: this.receivers.map(userObject => userObject.user),
 		}
@@ -75,28 +66,6 @@ export default {
 				this.$emit('update', v)
 			},
 		},
-
-		isValidQuery() {
-			return this.query?.trim() && this.query.length >= this.minSearchStringLength
-		},
-
-		options() {
-			if (this.isValidQuery) {
-				return this.suggestions
-			}
-			return []
-		},
-
-		noResultText() {
-			if (this.loading) {
-				return t('tables', 'Searching …')
-			}
-			return t('tables', 'No elements found.')
-		},
-
-		userId() {
-			return getCurrentUser().uid
-		},
 	},
 
 	methods: {
@@ -108,76 +77,11 @@ export default {
 			}
 		},
 
-		getShareTypes() {
-			const types = []
-			if (this.selectUsers) {
-				types.push(this.SHARE_TYPES.SHARE_TYPE_USER)
-			}
-			if (this.selectGroups) {
-				types.push(this.SHARE_TYPES.SHARE_TYPE_GROUP)
-			}
-			return types
+		filterOutUnwantedItems(list) {
+			list = this.filterOutCurrentUser(list)
+			return this.filterOutSelectedUsers(list)
 		},
 
-		getShareTypeString() {
-			if (this.selectUsers && !this.selectGroups) {
-				return 'User'
-			} else if (!this.selectUsers && this.selectGroups) {
-				return 'Group'
-			} else {
-				return 'User or group'
-			}
-		},
-
-		getPlaceholder() {
-			return t('tables', '{shareTypeString}...', { shareTypeString: this.getShareTypeString() })
-		},
-
-		async asyncFind(query) {
-			this.query = query.trim()
-			if (this.isValidQuery) {
-				this.loading = true
-				await this.debounceGetSuggestions(query)
-			}
-		},
-
-		async getSuggestions(search) {
-			this.loading = true
-			const shareTypes = this.getShareTypes()
-			let shareTypeQueryString = ''
-			shareTypes.forEach(shareType => {
-				shareTypeQueryString += `&shareTypes[]=${shareType}`
-			})
-			const url = generateOcsUrl(`core/autocomplete/get?search=${search}${shareTypeQueryString}&limit=${this.maxAutocompleteResults}`)
-
-			try {
-				const res = await axios.get(url)
-				const rawSuggestions = res.data.ocs.data.map(autocompleteResult => {
-					return {
-						user: autocompleteResult.id,
-						displayName: autocompleteResult.label,
-						icon: autocompleteResult.icon,
-						isUser: autocompleteResult.source.startsWith('users'),
-						key: autocompleteResult.source + '-' + autocompleteResult.id,
-					}
-				})
-
-				this.suggestions = this.filterOutCurrentUser(rawSuggestions)
-				this.suggestions = this.filterOutSelectedUsers(this.suggestions)
-				this.loading = false
-			} catch (err) {
-				console.debug(err)
-				showError(t('tables', 'Failed to fetch {shareTypeString}', { shareTypeString: this.getShareTypeString().toLowerCase() }))
-			}
-		},
-
-		debounceGetSuggestions: debounce(function(...args) {
-			this.getSuggestions(...args)
-		}, 300),
-
-		filterOutCurrentUser(list) {
-			return list.filter((item) => !(item.isUser && item.user === getCurrentUser().uid))
-		},
 		filterOutSelectedUsers(list) {
 			return list.filter((item) => !(item.isUser && this.localSharees.includes(item.user)))
 		},
